Moved priority queue classes into own files.
